+++ api/CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fixed an issue where the Rooted mount installer would be selectable for users who are non-root
 - Fixed importing patch bundles from storage taking much longer than it should to import (larger patch bundles till take some time) https://github.com/Jman-Github/Universal-ReVanced-Manager/issues/66
 - Fixed the occasional issue where a patch profile without Universal Patches would claim it requires universal patches enabled in order to use the said profile
+- Fixed slow download speeds for remote patch bundles
 
 
 # Docs